[UPDATE] Organização da API no projeto
