+++ no_fluxo_backend/testes_exemplo.sh
@@ -3,7 +3,7 @@
 # Script de exemplo para testar a API de testes do NoFluxo Backend
 # Certifique-se de que o servidor está rodando em localhost:3000
 
-BASE_URL="http://localhost:3000/fluxograma"
+BASE_URL="http://localhost:3000/testes"
 
 echo "🧪 Iniciando testes da API NoFluxo Backend"
 echo "=========================================="
@@ -11,17 +11,17 @@ echo "=========================================="
 # Teste 1: Conexão com banco
 echo ""
 echo "1️⃣ Testando conexão com banco..."
-curl -s -X GET "$BASE_URL/teste_banco" | jq '.'
+curl -s -X GET "$BASE_URL/banco" | jq '.'
 
 # Teste 2: Busca de curso
 echo ""
 echo "2️⃣ Testando busca de curso..."
-curl -s -X GET "$BASE_URL/teste_curso?nome_curso=Engenharia%20de%20Software" | jq '.'
+curl -s -X GET "$BASE_URL/curso?nome_curso=Engenharia%20de%20Software" | jq '.'
 
 # Teste 3: Casamento de disciplinas
 echo ""
 echo "3️⃣ Testando casamento de disciplinas..."
-curl -s -X POST "$BASE_URL/teste_casamento" \
+curl -s -X POST "$BASE_URL/casamento" \
   -H "Content-Type: application/json" \
   -d '{
     "nome_curso": "Engenharia de Software",
@@ -55,7 +55,7 @@ curl -s -X POST "$BASE_URL/teste_casamento" \
 # Teste 4: Teste completo
 echo ""
 echo "4️⃣ Executando teste completo..."
-curl -s -X POST "$BASE_URL/teste_completo" \
+curl -s -X POST "$BASE_URL/completo" \
   -H "Content-Type: application/json" \
   -d '{
     "nome_curso": "Engenharia de Software",
